# docs/openapi.yaml
openapi: 3.1.0
info:
  title: VelesDB API
  description: High-performance vector database for AI applications. Supports semantic search, HNSW indexing, and multiple distance metrics. Authentication is optional — when API keys are configured via VELESDB_API_KEYS, all endpoints except /health and /ready require a valid Bearer token.
  contact:
    name: VelesDB Team
    url: https://github.com/cyberlife-coder/VelesDB
  license:
    name: VelesDB Core License 1.0
    url: https://github.com/cyberlife-coder/VelesDB/blob/main/LICENSE
  version: 6.0.0
servers:
- url: /
  description: Local server
paths:
...
  /collections/{name}/empty:
    get:
      tags:
      - collections
      summary: Check if a collection is empty.
      operationId: is_empty
      parameters:
      - name: name
        in: path
        description: Collection name
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Empty status
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /collections/{name}/points:
    post:
      tags:
      - points
      summary: Upsert points to a collection.
      operationId: upsert_points
      parameters:
      - name: name
        in: path
        description: Collection name
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertPointsRequest'
        required: true
      responses:
        '200':
          description: Points upserted
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Invalid request or batch too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 'A body the server cannot deserialise: a field of the wrong type, or JSON that does not match the request schema. Axum''s extractor answers before the handler runs, so the payload is plain text, not an ErrorResponse.'
          content:
            text/plain:
              schema:
                type: string
  /collections/{name}/points/delete:
    post:
      tags:
      - points
      summary: Deletes multiple points by ID in a single request.
      description: |-
        Accepts a JSON body with a list of point IDs. All IDs are passed to
        the underlying `Collection::delete(&[u64])` in one call, which is
        more efficient than individual deletions.

        Returns the number of points that were requested for deletion.
        Points that do not exist are silently skipped (idempotent delete).

        # Empty payload semantics

        `{ "ids": [] }` is treated as a successful no-op: the response is
        `200 OK` with `deleted_count = 0`. This matches Kubernetes-style
        idempotent batch APIs and lets callers send empty batches without
        special-casing on the client side.

        # Limits

        Batches larger than `MAX_BULK_DELETE_SIZE` (10000) are rejected with
        `400 BAD_REQUEST`.
      operationId: bulk_delete_points
      parameters:
      - name: name
        in: path
        description: Collection name
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkDeleteRequest'
        required: true
      responses:
        '200':
          description: Points deleted
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Batch too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 'A body the server cannot deserialise: a field of the wrong type, or JSON that does not match the request schema. Axum''s extractor answers before the handler runs, so the payload is plain text, not an ErrorResponse.'
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Delete failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /collections/{name}/points/raw:
    post:
      tags:
      - points
      summary: Bulk upsert points via the binary wire format.
      description: |-
        Accepts an `application/octet-stream` body in the VRB1 raw-bulk format
        (a 16-byte header — magic, count, dim, id_width — followed by tightly-packed
        `u64` ids and `f32` vectors). The vectors are forwarded to
        `Collection::upsert_bulk_from_raw` (zero per-point `Point` allocation).
        Payloads are not supported on this path.
      operationId: upsert_points_raw
      parameters:
      - name: name
        in: path
        description: Collection name
        required: true
        schema:
          type: string
      requestBody:
        description: 'VRB1 binary bulk format: header + packed u64 ids + packed f32 vectors'
        content:
          application/octet-stream:
            schema:
              type: string
        required: true
      responses:
        '200':
          description: Points upserted
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Malformed body or dimension mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /collections/{name}/points/scroll:
    post:
      tags:
      - points
      summary: Scroll through collection points with cursor-based pagination.
      operationId: scroll_points
      parameters:
      - name: name
        in: path
        description: Collection name
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScrollRequest'
        required: true
      responses:
        '200':
          description: Scroll batch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScrollResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 'A body the server cannot deserialise: a field of the wrong type, or JSON that does not match the request schema. Axum''s extractor answers before the handler runs, so the payload is plain text, not an ErrorResponse.'
          content:
            text/plain:
              schema:
                type: string
  /collections/{name}/points/stream:
    post:
      tags:
      - points
      summary: Stream upsert points using NDJSON.
      description: |-
        Accepts a `application/x-ndjson` body. Each line is one JSON-encoded point,
        whose `id` is an unsigned 64-bit integer: a line that is not blank and does
        not parse as one, a string `id` included, is counted in the response's
        `malformed` field.
        Points are accumulated into micro-batches and flushed via `upsert_bulk`.

        The response body includes a `network_errors` field: a non-zero value means
        the HTTP body stream was truncated (e.g., client disconnect or proxy error),
        and the server may have received fewer points than the client sent (M-7).
      operationId: stream_upsert_points
      parameters:
      - name: name
        in: path
        description: Collection name
        required: true
        schema:
          type: string
      requestBody:
        description: NDJSON stream with one point per line
        content:
          application/x-ndjson:
            schema:
              type: string
        required: true
      responses:
        '200':
          description: Stream processed
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /collections/{name}/points/{id}:
    get:
      tags:
      - points
      summary: Get a point by ID.
      operationId: get_point
      parameters:
      - name: name
        in: path
        description: Collection name
        required: true
        schema:
          type: string
      - name: id
        in: path
        description: Point ID (u64 as a string; precision-safe above 2^53-1)
        required: true
        schema:
          type: string
          pattern: ^[0-9]+$
      responses:
        '200':
          description: Point found
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Point or collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
      - points
      summary: Delete a point by ID.
      operationId: delete_point
      parameters:
      - name: name
        in: path
        description: Collection name
        required: true
        schema:
          type: string
      - name: id
        in: path
        description: Point ID (u64 as a string; precision-safe above 2^53-1)
        required: true
        schema:
          type: string
          pattern: ^[0-9]+$
      responses:
        '200':
          description: Point deleted
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Point or collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /collections/{name}/points/{id}/relations:
    get:
      tags:
      - graph
      summary: List outgoing relation edges for a point.
      operationId: get_point_relations
      parameters:
      - name: name
        in: path
        description: Collection name
        required: true
        schema:
          type: string
      - name: id
        in: path
        description: Point ID (u64 as a string; precision-safe above 2^53-1)
        required: true
        schema:
          type: string
          pattern: ^[0-9]+$
      responses:
        '200':
          description: Outgoing relations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelationsResponse'
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /collections/{name}/points/{id}/ttl:
    patch:
      tags:
      - points
      summary: Set (or refresh) the durable TTL of a point.
      description: |-
        Persists `_veles_expires_at` in the point's payload so the expiry
        survives a restart. A `ttl_seconds` of `0` expires the point immediately.
        Expired points are excluded from all read surfaces (search/get/scroll/query);
        refreshing an expired point returns 404; storage is reclaimed lazily.
      operationId: set_point_ttl
      parameters:
      - name: name
        in: path
        description: Collection name
        required: true
        schema:
          type: string
      - name: id
        in: path
        description: Point ID (u64 as a string; precision-safe above 2^53-1)
        required: true
        schema:
          type: string
          pattern: ^[0-9]+$
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetTtlRequest'
        required: true
      responses:
        '204':
          description: TTL set successfully
        '400':
          description: Non-object payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Collection or point not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 'A body the server cannot deserialise: a field of the wrong type, or JSON that does not match the request schema. Axum''s extractor answers before the handler runs, so the payload is plain text, not an ErrorResponse.'
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /collections/{name}/relations:
    post:
      tags:
      - graph
      summary: Create a relation edge between two points in a collection.
      description: |-
        Works on vector, graph, and metadata collections alike.
        The edge ID is auto-assigned; the response body carries the allocated value.
      operationId: relate_points
      parameters:
      - name: name
        in: path
        description: Collection name
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RelateRequest'
        required: true
      responses:
        '201':
          description: Relation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelateResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Collection not found, or source/target point has no stored payload (VELES-022 NodeNotFound)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 'A body the server cannot deserialise: a field of the wrong type, or JSON that does not match the request schema. Axum''s extractor answers before the handler runs, so the payload is plain text, not an ErrorResponse.'
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /collections/{name}/relations/{edge_id}:
    delete:
      tags:
      - graph
      summary: Remove a relation edge by ID.
      operationId: unrelate_points
      parameters:
      - name: name
        in: path
        description: Collection name
        required: true
        schema:
          type: string
      - name: edge_id
        in: path
        description: Edge ID to remove (u64 as a string; precision-safe above 2^53-1)
        required: true
        schema:
          type: string
          pattern: ^[0-9]+$
      responses:
        '204':
          description: Relation removed
        '404':
          description: Collection or edge not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    BulkDeleteRequest:
      type: object
      description: Request body for bulk point deletion.
      required:
      - ids
      properties:
        ids:
          type: array
          items:
            oneOf:
            - type: integer
              format: int64
              minimum: 0
            - type: string
              pattern: ^[0-9]+$
          description: Point IDs. Each accepts a JSON integer (native form) or a string; use a string for u64 values above 2^53-1 to avoid JavaScript precision loss.
...
    ErrorResponse:
      type: object
      description: Error response.
      required:
      - error
      properties:
        code:
          type:
          - string
          - 'null'
          description: Structured error code (e.g., `"VELES-004"` for dimension mismatch).
        error:
          type: string
          description: Error message.
...
    ScrollRequest:
      type: object
      description: Request body for the scroll endpoint.
      properties:
        batch_size:
          type: integer
          format: int32
          description: Maximum points per batch (1–10 000, default 100).
          minimum: 0
        cursor:
          oneOf:
          - type: integer
            format: int64
            minimum: 0
          - type: string
            pattern: ^[0-9]+$
          - type: 'null'
          description: Resume after this point ID (exclusive). Omit or send null to start from the beginning. Accepts a JSON integer or a precision-safe string.
        filter:
          description: Optional filter expression.
...
    UpsertPointsRequest:
      type: object
      description: Request to upsert points.
      required:
      - points
      properties:
        points:
          type: array
          items:
            $ref: '#/components/schemas/PointRequest'
          description: Points to upsert.
...
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
security:
- bearer_auth: []
...